changes in folder structure and modularised scss
