--- src/components/ArtBanner/ArtBanner.scss
@@ -1,44 +1,0 @@
-.artBanner{
-    height:  calc(100vh - var(--nav-height));
-    margin: var(--nav-height) 0;
-    background-image: url("./../../../public/assets/art-image.webp") ;
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-position-y: center;
-    background-clip: border-box;
-    width: 100%;
-    // padding: var(--padding);
-    color: black;
-
-    display: flex;
-    align-items: flex-end;
-    justify-content: center;
-}
-
-.artBanner .contents{
-    width: 90%;
-    margin-inline: auto;
-    height: max-content;
-    border-radius: 5px;
-    h1,p, .button-wrapper{
-        padding: var(--padding-small) 0;
-        max-width: 35em;
-    }
-}
-
-.button-wrapper{
-    width: 100%;
-
-}
-
-@media only screen and (max-width: 480px){
-    .artBanner{
-        padding: 16px;
-
-        background-position:  30% ;
-        .contents{
-            width: 100%;
-        }
-    }
-}
-
