--- src/components/PhotoBanner/PhotoBanner.scss
@@ -1,61 +1,0 @@
-.photoBanner{
-    height: calc(100vh - var(--nav-height));
-    background-image: url("../../../public/assets/PhotoBanner.webp"), linear-gradient(rgba(155, 155, 155, 0.749), rgba(0, 0, 0, 0.864));
-    
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-position: center 75%;
-    background-clip: border-box;
-    width: 100%;
-    display: flex;
-    justify-content: flex-start;
-    align-items: flex-end;
-    color: white;
-}
-.photoBanner .contents{
-    width: 90%;
-    margin-inline: auto;
-    position: relative;
-    border-radius: 5px;
-    animation: slide-up 0.7s ease-in forwards;
-    h1,p, .button-wrapper{
-        padding: var(--padding-small) 0;
-        max-width: 35em;
-    }
-}
-@keyframes slide-up {
-    0%{
-        bottom: -4vh;
-    }
-    100%{
-        bottom: 4vh;
-    }
-}
-
-p{
-    line-height: 24px;
-}
-.button-wrapper{
-    width: 100%;
-
-}
-
-
-@media only screen and (max-width: 480px) {
-    .photoBanner{
-        padding: 16px;
-        background-image: 
-        linear-gradient(to top, #00000069, transparent,transparent),
-        url("../../../public/assets/PhotoBanner.webp"),
-        linear-gradient(to top,rgb(0 0 0 / 89%),rgb(68 68 68 / 75%), transparent);
-        
-        .contents{
-            width: 100%;
-            h1,p,.button-wrapper{
-                max-width: 100%;
-            }
-        }
-    } 
-    
-}
-
