+++ src/pages/HomePage/Home.css
@@ -1,0 +1,3 @@
+.container{
+    grid-area: content;
+}
